wanted to commit this stuff before I accidentally deleted it Should add min-max capability to vjAnalogs,
not tested yet,  I'll do that once juggler bug is gone and building like it used to without messing with
experimental stl libs and such.

config files are updated to this change also...


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler_1.0@2575 08b38cba-cd3b-11de-854e-f91c5b6e4272

+++ Input/vjInput/vjDummyAnalog.h
@@ -69,7 +69,7 @@ class vjDummyAnalog : public vjAnalog {
 	char* getDeviceName() { return "vjDummyAnalog"; }
 	
 	//: Get the analog data
-   int getAnalogData (int d) { return 0; }
+   float getAnalogData (int d) { return 0; }
 };
 
 #endif
